requirements.txt
Django==4.0.6
celery==5.2.7
django-celery-beat==2.3.0
psycopg2==2.9.3
requests==2.28.1
beautifulsoup4==4.11.1
lxml==4.9.1
python-slugify==6.1.2
django-better-admin-arrayfield==1.4.2
